Refactor tei rerank
