+++ comps/reranks/deployment/docker_compose/rerank_tei.yaml
@@ -4,18 +4,23 @@
 version: "3.8"
 
 services:
-  tei_xeon_service:
+  tei_reranking_service:
     image: ghcr.io/huggingface/text-embeddings-inference:cpu-1.2
-    container_name: tei-xeon-server
+    container_name: tei-reranking-server
     ports:
       - "8808:80"
     volumes:
       - "./data:/data"
     shm_size: 1g
+    healthcheck:
+      test: ["CMD", "curl", "-f", "http://localhost:8808/health"]
+      interval: 10s
+      timeout: 6s
+      retries: 18
     command: --model-id ${RERANK_MODEL_ID} --hf-api-token ${HF_TOKEN}
   reranking:
-    image: opea/reranking-tei:latest
-    container_name: reranking-tei-xeon-server
+    image: opea/reranking:latest
+    container_name: reranking-tei-server
     ports:
       - "8000:8000"
     ipc: host
@@ -25,6 +30,9 @@ services:
       https_proxy: ${https_proxy}
       TEI_RERANKING_ENDPOINT: ${TEI_RERANKING_ENDPOINT}
       HF_TOKEN: ${HF_TOKEN}
+    depends_on:
+      tei_reranking_service:
+        condition: service_healthy      
     restart: unless-stopped
 
 networks:
