--- comps/reranks/deployment/docker_compose/videoqna.yaml
@@ -1,21 +1,0 @@
-# Copyright (C) 2024 Intel Corporation
-# SPDX-License-Identifier: Apache-2.0
-
-services:
-  reranking:
-    image: opea/reranking-videoqna:latest
-    container_name: reranking-videoqna-server
-    ports:
-      - "8000:8000"
-    ipc: host
-    environment:
-      no_proxy: ${no_proxy}
-      http_proxy: ${http_proxy}
-      https_proxy: ${https_proxy}
-      CHUNK_DURATION: ${CHUNK_DURATION}
-      FILE_SERVER_ENDPOINT: ${FILE_SERVER_ENDPOINT}
-    restart: unless-stopped
-
-networks:
-  default:
-    driver: bridge
